refactor: delete file and move its function to SignIn
